Banco NoSQL/Atividade Memcached/requirements.txt
Flask==2.2.2
python-memcached==1.59
requests==2.28.1